Refactor input sanitization: move utility functions to middlewares and update imports
